net: route fetch responses in the net main loop and expose a safe URLSession API
Fetch response routing moves out of the net backends into the net process's
main loop. NetworkBackend::http_network_or_cache_fetch now takes a crossbeam
reply channel instead of a reply_to recipient: backends deliver
(request_id, result) on it — synchronously for tokio, from the URLSession
completion handler for url_session — and the main loop select!s over the
request receiver and the reply receiver, looks up the request's stored
reply_to, and routes the outcome to the one-off content command channel
(ResponseRecipient::ContentProcess) or the persistent net→UA channel
(ResponseRecipient::UserAgent). ResponseRecipient::UserAgent no longer
embeds a response sender; the net process replies on its own bootstrap
connection's sender. IpcConnection drops the incoming_sender field (a plain
sender/receiver pair again), the bootstrap no longer ships a duplicate
child→parent sender, and the user agent no longer retains or embeds a reply
sender.

url_session_sys now offers only a safe API: UrlSession::new, UrlSession::fetch,
and FetchResponse are its public surface, with the raw FFI declarations
(fw_url_session_*) in a crate-private ffi module. The duplicate safe wrapper
crate net/url_session is deleted and the URLSession backend imports
url_session_sys directly, so no call site outside the sys crate writes
unsafe. The C wrapper's error path no longer passes a NULL error pointer
(which the trampoline reads as success), and the URLCache=nil decision and
the missing cancellation path are documented in code.

Prune the per-algorithm flow walkthroughs from the content/src/html and
user_agent readmes ("Iframe navigation flow", "Window.open flow", "Posting
messages", "The rules for choosing a navigable", "Document creation: two
directions", "Window.open steps", "Document creation for new traversables"):
the per-step split is carried by the // Step N: comments in the code. Keep
general guidance and remaining-gap tracking; trim the opener-tracking section
to its durable facts.
