+++ Cargo.toml
@@ -53,7 +53,6 @@ members = [
     "js_engine_macros",
     "media",
     "net",
-    "net/url_session",
     "net/url_session_sys",
     "user_agent",
     "verification",
