+++ net/Cargo.toml
@@ -11,7 +11,7 @@ default = []
 # only) backend and reqwest is always compiled (see the target-specific
 # dependency tables); on Apple it is opt-in.
 tokio = ["dep:reqwest"]
-# Apple URLSession backend (macOS/iOS only). Marker feature: the url_session
+# Apple URLSession backend (macOS/iOS only). Marker feature: the url_session_sys
 # crate is always compiled on Apple (see the target-specific dependency
 # tables), so this flag just selects the backend when the default
 # (`url_session_default` from build.rs) is not what is wanted.
@@ -27,6 +27,7 @@ path = "src/bin/net_process.rs"
 [dependencies]
 ipc = { path = "../ipc" }
 ipc_messages = { path = "../ipc_messages" }
+crossbeam-channel = "0.5"
 mime_guess = "2"
 log = "0.4"
 env_logger = "0.11"
@@ -43,5 +44,5 @@ reqwest = { version = "0.12.28", default-features = false, features = ["blocking
 [target.'cfg(any(target_os = "macos", target_os = "ios"))'.dependencies]
 reqwest = { version = "0.12.28", default-features = false, features = ["blocking", "rustls-tls"], optional = true }
 # The URLSession backend is always available on Apple: the default backend
-# (url_session_default from build.rs) and the feature share the same crate.
-url_session = { path = "url_session" }
+# (url_session_default from build.rs) and the feature share the same crate (url_session_sys).
+url_session_sys = { path = "url_session_sys" }
