+++ net/README.md
@@ -25,7 +25,7 @@ the `url_session_default` cfg when no backend feature is enabled, so a
 plain `cargo build` compiles no reqwest/tokio stack there. On non-Apple
 platforms the tokio backend is the only option and is always compiled. The
 tokio backend on Apple is opt-in via `--features tokio` (the URLSession
-feature flag itself is a marker — the `url_session` crate is always
+feature flag itself is a marker — the `url_session_sys` crate is always
 compiled on Apple).
 
 The root build prebuilds `formal-web-net` with the platform default backend
@@ -40,26 +40,28 @@ the tokio backend stores one reqwest client per key, the URLSession backend
 stores one session (with no shared cache) per key. Sessions are created on
 first use and reused.
 
-### Response routing lives in the backends
+### Response routing lives in the main loop
 
-`route_response` (`backend/mod.rs`) delivers the fetch outcome to the
-`reply_to` recipient, which carries the response sender in both cases — the
-content process's command sender, or the user agent's response channel. The
-UA reply channel is modelled on the fetch spec's
-[parallel queue](https://fetch.spec.whatwg.org/#fetch-useparallelqueue):
-responses are delivered onto it. Because routing lives in the backend, the
-response may be sent at any time: the tokio backend sends it before its
-fetch method returns, the URLSession backend sends it from the data task's
-completion handler (a background queue). The UA's sender end of the
-net→UA response channel is retained at extension launch by the ipc crate
-(`IpcConnection::incoming_sender`) and embedded into
-`ResponseRecipient::UserAgent` by `user_agent/src/fetch.rs`.
+The backends deliver fetch outcomes on a plain crossbeam reply channel: each
+`http_network_or_cache_fetch` call takes a clone of the process-wide
+`FetchReplySender` (`backend/mod.rs`) and sends `(request_id, result)` on it.
+The main loop (`main.rs`) stores the `reply_to` recipient of every in-flight
+request in a pending map, `select!`s over the request receiver and the reply
+receiver, and routes each delivered outcome to its recipient: the one-off
+content command channel embedded in `ResponseRecipient::ContentProcess`, or
+the persistent net→UA channel — the sender end of the net process's own
+bootstrap connection — for `ResponseRecipient::UserAgent`. Because routing
+lives in the main loop, a backend may deliver a reply at any time after its
+fetch method returns: the tokio backend sends it before returning, the
+URLSession backend sends it from the data task's completion handler (a
+background queue).
 
-## url_session and url_session_sys
+## url_session_sys
 
-- `net/url_session` — safe Apple URLSession client: one session with no
-  shared cache, fetch completions delivered on a background queue.
-- `net/url_session_sys` — raw Objective-C bindings: a C/ObjC wrapper
-  (`src/url_session_wrapper.m`, compiled via `cc`) exposing a small C API
-  (`fw_url_session_create/fetch/release`), following the `ipc/xpc-sys`
-  pattern. Apple targets only.
+`net/url_session_sys` is the Apple URLSession client: a safe Rust API (one
+session with no shared cache, fetch completions delivered on a background
+queue) over a private Objective-C FFI layer. The raw FFI declarations
+(`src/ffi.rs`) and the C/ObjC wrapper (`src/url_session_wrapper.m`, compiled
+via `cc`, exposing the small C API `fw_url_session_create/fetch/release`)
+are crate-private; the public surface — `UrlSession::new`, `UrlSession::fetch`,
+and the `FetchResponse` type — is safe. Apple targets only.
