+++ net/src/backend/url_session.rs
@@ -5,11 +5,10 @@ use std::collections::HashMap;
 use std::collections::hash_map::Entry;
 
 use ipc_messages::content::{FetchRequest, FetchResponse};
-use ipc_messages::network::ResponseRecipient;
-use url_session::UrlSession;
+use url_session_sys::UrlSession;
 use uuid::Uuid;
 
-use super::{NetworkBackend, NetworkPartitionKey, handle_local_schemes, route_response};
+use super::{FetchReplySender, NetworkBackend, NetworkPartitionKey, handle_local_schemes};
 
 /// <https://fetch.spec.whatwg.org/#http-network-fetch>
 pub struct UrlSessionBackend {
@@ -47,25 +46,37 @@ impl NetworkBackend for UrlSessionBackend {
         key: NetworkPartitionKey,
         request_id: Uuid,
         request: &FetchRequest,
-        reply_to: ResponseRecipient,
+        reply_sender: FetchReplySender,
     ) -> Result<(), String> {
         // Local schemes never touch the transport.
         match handle_local_schemes(request) {
-            Ok(Some(response)) => return route_response(request_id, reply_to, Ok(response)),
-            Err(error) => return route_response(request_id, reply_to, Err(error)),
+            Ok(Some(response)) => {
+                return reply_sender
+                    .send((request_id, Ok(response)))
+                    .map_err(|error| format!("failed to send fetch reply: {error}"));
+            }
+            Err(error) => {
+                return reply_sender
+                    .send((request_id, Err(error)))
+                    .map_err(|error| format!("failed to send fetch reply: {error}"));
+            }
             Ok(None) => {}
         }
 
         let session = match self.session_for(key) {
             Ok(session) => session,
-            Err(error) => return route_response(request_id, reply_to, Err(error)),
+            Err(error) => {
+                return reply_sender
+                    .send((request_id, Err(error)))
+                    .map_err(|error| format!("failed to send fetch reply: {error}"));
+            }
         };
 
         let method = request.method.clone();
         let url = request.url.clone();
         let body = request.body.clone();
         let body_bytes = (!body.is_empty()).then_some(body.as_bytes());
-        let completion_reply_to = reply_to.clone();
+        let completion_reply_sender = reply_sender.clone();
         if let Err(error) = session.fetch(&method, &url, body_bytes, move |result| {
             let result = result.map(|response| FetchResponse {
                 final_url: response.final_url,
@@ -73,14 +84,16 @@ impl NetworkBackend for UrlSessionBackend {
                 content_type: response.content_type,
                 body: response.body,
             });
-            if let Err(route_error) = route_response(request_id, completion_reply_to, result) {
-                log::error!("failed to route URLSession response: {route_error}");
+            if let Err(send_error) = completion_reply_sender.send((request_id, result)) {
+                log::error!("failed to route URLSession response: {send_error}");
             }
         }) {
             // The task could not be started; the completion callback was not
             // invoked, so deliver the failure to the caller now.
             let result = Err(format!("failed to start URLSession fetch: {error}"));
-            return route_response(request_id, reply_to, result);
+            return reply_sender
+                .send((request_id, result))
+                .map_err(|error| format!("failed to send fetch reply: {error}"));
         }
         Ok(())
     }
