--- net/url_session/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "url_session"
-version = "0.1.0"
-edition = "2024"
-publish = false
-
-[lib]
-path = "src/lib.rs"
-
-[dependencies]
-url_session_sys = { path = "../url_session_sys" }
